netbox_plugin_gpon/pyproject.toml
[tool.poetry]
name = "netbox-plugin-gpon"
version = "0.1.0"
description = ""
homepage = "https://github.com/MajesticFalcon/netbox-plugin-gpon"
repository = "https://github.com/MajesticFalcon/netbox-plugin-gpon"
authors = ["sutley"]

packages = [
    { include = "netbox_plugin_gpon" },
]
[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^20.8b1"
invoke = "^1.5.0"
pylint = "^2.6.0"
pylint-django = "^2.4.2"
pydocstyle = "^5.1.1"
yamllint = "^1.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
